fix: enable Nostalgia pack page
